Add a "relational" variant of the pearl.
In this version the type of elements in the list, tree and intermediate datastructures are replaced by a graph (or proof relevant relation) on an abstract base type.

By parametricity this is supposed to force the order of elements to be preserved, avoiding the need for the boring proof in `Proof.v`.